+++ test-crate/tests/macro-errors-single.rs
@@ -1,7 +1,0 @@
-extern crate trybuild;
-
-#[test]
-fn ui() {
-    let t = trybuild::TestCases::new();
-    t.compile_fail("tests/ui-single/*.rs");
-}