+++ test-crate/tests/ui-single/span_single.rs
@@ -1,7 +1,0 @@
-extern crate test_crate;
-
-use test_crate::make_fn;
-
-make_fn!(span_single);
-
-fn main() {}