+++ test-crate/tests/ui/call_site_many.rs
@@ -1,0 +1,7 @@
+extern crate test_crate;
+
+use test_crate::make_fn;
+
+make_fn!(call_site_many);
+
+fn main() {}