+++ test-crate/tests/ui/result_expect.rs
@@ -1,0 +1,7 @@
+extern crate test_crate;
+
+use test_crate::make_fn;
+
+make_fn!(result_expect);
+
+fn main() {}