+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.2.5 (2019-09-02)
+* Introduce support for dummy implementations via `dummy::set_dummy`
+* `multi::*` is now deprecated, will be completely rewritten in v0.3
+
 # v0.2.0 (2019-08-15)
 
 ## Breaking changes